error message

+++ site/html/logic/delete_user.php
@@ -23,13 +23,18 @@ try {
     if (isset($file_db)) {
         // get user email passed through url
         $email = $_GET['email'];
-        echo $email;
-        // delete user from db
-        $sql = $file_db->prepare("DELETE FROM users WHERE email=:email");
-        $sql->bindParam('email', $email);
-        $sql->execute();
-        //redirect to users management
-        header('Location: ../view/users.php');
+
+        if ($email != $_SESSION['email']) {
+            // delete user from db
+            $sql = $file_db->prepare("DELETE FROM users WHERE email=:email");
+            $sql->bindParam('email', $email);
+            $sql->execute();
+            //redirect to users management
+            header('Location: ../view/users.php');
+        } else {
+            $_SESSION['error'] = "You can't delete yourself";
+            header('Location: ../view/users.php');
+        }
     }
 } catch (PDOException $e) {
     // Close file database.sqlite connection
